# Resolved conflicts:
#	Jint/Jint.csproj

+++ Jint/Constraints/ConstraintsOptionsExtensions.cs
@@ -20,17 +20,23 @@ public static class ConstraintsOptionsExtensions
         return options;
     }
 
+    /// <summary>
+    /// Sets constraint based on memory usage in bytes.
+    /// </summary>
     public static Options LimitMemory(this Options options, long memoryLimit)
     {
         options.WithoutConstraint(x => x is MemoryLimitConstraint);
 
-        if (memoryLimit > 0 && memoryLimit < int.MaxValue)
+        if (memoryLimit > 0 && memoryLimit < long.MaxValue)
         {
             options.Constraint(new MemoryLimitConstraint(memoryLimit));
         }
         return options;
     }
 
+    /// <summary>
+    /// Sets constraint based on fixed time interval.
+    /// </summary>
     public static Options TimeoutInterval(this Options options, TimeSpan timeoutInterval)
     {
         if (timeoutInterval > TimeSpan.Zero && timeoutInterval < TimeSpan.MaxValue)
@@ -40,6 +46,9 @@ public static class ConstraintsOptionsExtensions
         return options;
     }
 
+    /// <summary>
+    /// Sets cancellation token to be observed. NOTE that this can be unreliable/imprecise on full framework due to timer logic.
+    /// </summary>
     public static Options CancellationToken(this Options options, CancellationToken cancellationToken)
     {
         options.WithoutConstraint(x => x is CancellationConstraint);
